requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
requests
pypdf
nltk
numpy

